ci: fix daily-reset (#18121)
I wasn't sure what exactly was causing the issue so I removed everything
I thought might be:

- Added quotes to the GET_TARGETS heredoc.
- Remove the jq invocation at the end of the heredoc.
- Removed the subshells in crontab.

This now seems to work, though the (limited) output I'm seeing on the
machine is still not quite what I expected. Some of the changes may not
have been strictly needed, but given how slow the turnaround for this is
I'd prefer not to try and see which ones.