fix blackduck script update (#18557)
This job runs on both `main` and `main-2.x`, and generates the same
branch name. This results in only the first of the two that runs
generating a PR.

With this change the branch name will be different so when an update is
needed we'll update both on the same day.