JSON API: when a new party makes a query, only load the cache from ACS if there are no pre-existing contracts visible to that party (#18636)
Part of https://digitalasset.atlassian.net/browse/ETX-421

This implementation of the fix attempts to minimise changes along two axes.

The first is to the performance characteristics of the JSON API service. That's why we only alter the behaviour to use transactions instead of ACS in the specific case where the bug arises, i.e. when a party issues a query for a template for the first time **and** there are pre-existing contracts of this template visible to that party.

The second is the structure of the code, simple diffs will make merges easier when we backport this fix to version 2.3, 2.7 and 2.8

The green from commit 0822650 demonstrate this working on Oracle
https://github.com/digital-asset/daml/runs/22315131743