[release] test faster releases
This is an attempt to test whether the configuration in #18812 is
sufficient to successfully create a release, as I try to find out the
minimal set of things we need to explicitly build.

Note that by its very nature the ad-hoc release we're creating here will
reference a transient, ephemeral commit.