JSON API: handle zombie contract using `oldestVisibleOffset` instead of `hasVisibleContracts` (#18866)
In the case of [zombie contracts](https://digitalasset.atlassian.net/browse/ETX-421?focusedCommentId=128170), we were previously populating the cache using getTransactions from `LedgerBegin`. If the JSON API cache has been wiped after a ledger prune, this operation will fail.

In order to handle this case, we need to avoid attempting to load all transactions since `LedgerBegin`, and instead load from the oldest offset that could possibly have lead to a zombie contract. This will be the minimum `last_offset` of any other party who shares contracts of the relevant template with our current querying part(y|ies). If the cache was cleared after a ledger prune, then any `last_offset` found in the cache must be an offset that occurred after the prune, and is thus safe to fetch from.