Metrics: initialise views before meters to re-enable histograms (#18963)
Before this change, the whole histogram building logic was broken since we upgraded telemetry. Now,
we initialize the view before the meters. As part of this change, I also move the documentation
into the meter definition such that the metrics can include a description in otel, and finally I've improved
the filtering such that we now also filter by qualification, which means that a client will not be
overwhelmed with 400 unnecessary metrics.

Part of https://github.com/DACH-NY/canton/issues/17917