JSON API Upgrades: ETX-500 require fully qualified template ids (#19149)
- update all instances of ContractTypeId.OptionalPkg to ContractTypeId.RequiredPkg
- removed ContractTypeId.OptionalPkg and ContractTypeId.NoPkg type aliases entirely
- update tests to use fully qualified template ids, i.e. including a package reference
- update integration tests which do json-api queries to include a package reference in template ids

We still need to distinguish package name vs package id at the point where we resolve contract type ids, so we keep the logic that we used on missing package id, and do the same thing in the case of a package name. This logic deals with the fact that new packages can get uploaded in the background, and so we need to be able to treat a package name as a dynamically late-bound mapping to specific package.