HOTFIX: remove `sdk/` prefix from paths when running `./fmt.sh --diff` (#19237)
As a developer, we now run `./fmt.sh --diff` from within the `sdk/` dir, but git still returns a list of changed files with a sdk prefix. This change removes the sdk/ prefix from those paths.

Incidentally, we were previously referring to `${changed_files[@]}` as though it were a bash array, but it was a regular scalar variable. This makes it actually an array, which allows us to use bash builtin syntax to strip the prefixes.