JSON API: etx-500 add type information into the package ids of template ids (#19211)
A template/interface id (`ContractTypeId`) consists of three parts: package, module and entity.

So far the package portion has been modeled with a `String`, where that string could represent either a package name or a package id, depending on whether it started with a `"#"` prefix or not, respectively. However there are parts of the code, namely those that deal with the db cache, that are designed to work with contract type ids whose package portion is only a *package id*. The type system was not helping us to ensure this so far.

Fortunately the `ContractTypeId` class is parameterised with the type of the package, so this PR replaces all existing instantiations which used `String` packages, with either [PackageId](https://github.com/digital-asset/daml/blob/7ffc096fa857700fe510c41f299c7ce06f5bbb4f/sdk/daml-lf/data/src/main/scala/com/digitalasset/daml/lf/data/Ref.scala#L37) or [PackageRef](https://github.com/digital-asset/daml/blob/7ffc096fa857700fe510c41f299c7ce06f5bbb4f/sdk/daml-lf/data/src/main/scala/com/digitalasset/daml/lf/data/Ref.scala#L292) (which is a sum type which can contain either a `PackageId` or a `PackageName`).

The `ContractTypeRef` type contains all the package info we have about a contract type id. Its `original` field now returns a `ContractTypeId[PackageRef]`, whereas its `latestPkgId` returns a `ContractTypeId[PackageId]`. This helps us to ensure that we are passing around the right kind of package specifiers to the parts of the code that need one or the other.

There are some type aliases used, that may be useful to help read and understand the changes, which tend to look like
```
   type RequiredPkg = CtId[Ref.PackageRef]
   type RequiredPkgId = CtId[Ref.PackageId]
```
i.e. if it ends with Pkg it uses a `PackgeRef` and if it ends with PkgId uses a `PackageId`. 
Note also that the `Resolved` terminology is an orthogonal concern which refers to whether this id has been determined to belong to a template vs interface (i.e. it extends `Definite`), or whether that is not yet known.