prevent ValueTranslation usage on non serializable data (#20011)
This is not absolutely necessary. However taking into account several tests were actually using non serializable data, I think this sanity check can be useful to prevent misusing the translator.

Note we discovered a small bug in the compiler that was not running the Serializability inference when type-checking was disable. (Thanks for @samuel-williams-da to quickly fine and propose the fix included in this PR)

We also rename TypeF in SerializableTypeF to make it clear the TypeDesctructor handles only serializable types.

---------

Co-authored-by: Samuel Williams <samuel.williams@digitalasset.com>