Update `MissingPackage` error to allow reporting by package name (#19972)
In the participant, the handling of a command which contains a template id whose package id is not known should be the same, irrespective of whether the template id specified the package with a package id or a package name. From that point of view, it makes sense to handle both cases with the same error.

If the `MissingPackage` type contained a `PackageRef` rather than a `PackageId` it could make this more uniform, and cases which did need to distinguish could still pattern match on the subtype of `PackageRef`, as needed.