JSON API: fail loudly if cache is newer than ledger (#20026)
Intended to make it more obvious when an existing cache database is re-used with a new ledger.

Previously we used `start >= end` to decide that our cache is up-to-date and we silently fetched nothing. This is a valid decision for the `start == end` case, but we should not expect `start > end` under normal use. This is something we can see if they have started json api with a cache that was populated against a different ledger history, in which case we have no recourse other than :boom: 

Inspired by
https://support.digitalasset.com/s/case/500Pq00000LS2dwIAD/inconsistence-querying-jsonapi