[Doc] Rewrite the `Test Daml contracts` section in daml tuto (#21259)
This is part of DACH-NY/docs-website#313

It contains the following changes to the Test Daml contracts chapter in Get started with smart contract development tutorial:

   - Remove section about Contention errors
   - Create new daml-intro-12 template, which is a self-contained, multi-package version of daml-intro-7 and daml-intro-9 combined. Templates and tests are defined in separate packages.
   - Add a Multi-package project structure section
   - Remove mentions of Daml Navigator and Daml REPL
   - Don't explain how to run script against the sandbox or a running Canton ledger.
   - Trim section about test coverage, add link to corresponding how-to section (which is currently missing)
   - Remove daml-intro-12-part1 and daml-intro-12-part2 templates

In order to test the new daml-intro-12 template, we added a daml_multi_package_test rule in Bazel. It builds the entire multi-package project, and runs the tests in each sub-package.