Use real package metadata in script service (#21393)
This PR is a follow-up of #21351, to fix the bug described in #21324 (comment).

In the script service we used to create a dummy package metadata for the self package. This causes the queryInterface to fail resolving the interface view of an upgraded template in the self package. To fix it, the daml assistant reads the real package metadata from the daml.yaml and sends it to the script service through the UpdatePackages request.

As a side improvement, when we create the context of an external package, to run the scripts in that package, we reuse the loaded package, instead of loading its modules separatly. This ensures we don't create two packages with the same signature but different package ids. The package metadata is used to determine the real package id of the loaded package that contains the scripts to run.


* Remove unused code in ScriptServiceClient

* Fix package vetting test

* Change diagnostics in PackageVetting test

* apply review