Better ValidationFailed error (#22011)
Resolve longstanding TODO to add more meta information to the upgrade
validation error.

Because resources in grpc errors do not represent fields (like "original
signatories" or "recomputed signatories") but instead represent field
types (like "parties"), the daml-script parser of these errors has to
rely on the order of the resources to reconstruct the error datatype
from the grp response. This makes things awkward when some resources are
optional. For instance, until this PR, we had to handle two cases for
ValidationFailed: the case where the key was present, and the case where
it wasn't. Now that we're having two keys in the error, we would need to
handle 4 cases: no key present, both keys present, only first key
present, only second key present. In order to avoid that, this PR
introduces nullable resources, which can be either a value or NULL. It
is still pretty awkward, but at least less error prone.