Revert "Open up IsParties instances (#21390)" (#22058)
This reverts commit 3a72b882b9560d17f69067d10baa2f658561f7ff.

The commit this reverts changed the instances for IsParties in the
stdlib.
Unfortunately, when using data-deps from before this change, we now run
into instance collisions, between
`IsParties a => IsParties [a]` and `IsParties [Party`, for example.
To do this properly we would need to keep the original instances, and
apply the correct overlap pragmas (and ensure said pragmas survive
data-deps).
Or we'd need bespoke logic in data-deps to deal with the older
instances.
We don't want to risk this for 3.4, given CN isn't using the new
instances yet.