Reject trailing nones for v11 and v12 contracts in value import (#22064)
This PRs adds a parameter to `speedy.ValueTranslator` which allows for
the rejection of trailing Nones in records. We reject trailing Nones for
v11 and v12 contracts and allow it for v10 contracts, using
HashingMethod as a proxy for the contract ID version.

This PR also adds a check that numerical value have the expected scale.