add enrichContractWithPackages method to the enricher (#22066)
For external signing, we want to reject transactions in which the same
input contract is upgraded using different packages which enrich the
contract to different values. This can happen when using v10 or v11
contracts because authentication of these contracts won't catch a
relabeling.

In order to achieve this, we expose to canton a new
`enrichContractWithPackages` function which takes a fat contract
instance and a non-empty set of packages, checks that they all hash
(using the TypedNormalForm hashing method) to the same value, and then
returns the enrichment of the minimal package ID.