Detect inconsistent package names in fat contract instances (#22073)
Fat contract instances (FCIs) declare the name of the package that
created them. When we exercise dynamic (by interface) choice on a FCI,
we use that package name to look up the preferred package with that
name. But for static exercises, we ignore that name altogether: instead,
we use the daml-model provided package ID as the target package ID. As
part of the upgrade to that package ID, we authenticate the contract,
which checks that the package name of the target package matches that of
the creating package.

But what we do *not* check in that case, is that the package name of the
target package matches that of the FCI. It is not really an issue, as
this package name is unused. But it means we accept ill-formed FCIs,
which is weird and may come back to bite us.

This PR fixes that.