rm non-stable daml-prim from list stable pkgs (#22101)
This fixes the bug described below. To debug this, we extended
`EngineValidatePackagesTest` with a way to run `validateDar` on a .dar,
which was the conflicting dar during debugging. To not check in any dars
we have changed it to `script-test-v2.dev.dar`, but it must be replaced
by a better dar later.

The bug:

1. The dar was compiled using canton 3.3, which means it does not have
explicit imports (any dar compiled on lf 2.1, even on 3.4, would have
also triggered this bug)
2. When generating the imports field, based on directdeps, the
_hardcoded_ set of stable packages is diffed away, which until this PR
erroneously included daml-prim
3. When comparing against the set of _included_ packages however, the
_generated_ set of stable packages, which does _not_ include daml-prim
is diffed away
4. Therefore, an error was reported: daml-prim is included, but not
mentioned