mv  daml-lf/engine tests to lf 2.2 (#22146)
Convert tests in //daml-lf/engine:all to use 2.2 instead of 2.1. The
rest will come in a separate PR.

To obtain these fixes, we set the decoder to crash upon encountering
2.1. We then ran all tests and whitelisted the ones that already test
both 2.1 and 2.2 (and possibly 2.dev), and fix the rest, until no more
remain. This is the batch of fixed tests limited to //daml-lf/engine:al