[backport] Implement hash-based authentication in the IDE ledger  (#22202)
Backport of #22184. Pure cherry-pick, no changes.

In the IDE ledger, suffixes the contract IDs of local contracts with the
TypedNormalForm hash of their create argument. In the
`ResultNeedContract` callback for authenticating contracts given their
hash, compares the provided hash to the suffix of the contract ID.

The scenarios enabled by this change were already tested but these tests
were marked as broken on the IDE ledger. They now pass.

Also removes the `value: SValue` field from `IdeLedgerRunner.Commit`: it
is currently unused but if one were to use it, it would be wrong because
the contract IDs this SValue contains are not suffixed.