Blackduck CI job: Fix the NOTICES file PR creation (#22219)
I have tried out this change by forcing the job to run on this PR (the
commits doing that has since been reverted). One thing I had to do for
it to work was to set the base branch to `main` explicitly because when
running on PRs, `$(Build.SourceBranchName)` is not `main`. I believe
that when this job run with the daily compat then it will be `main`.