[3.4][Backport] Use scalapb options from Canton (#22225)
Backport of:

- https://github.com/digital-asset/daml/pull/22215 

[This PR](https://github.com/DACH-NY/canton/pull/28047) is failing
because the scalapb generated on the canton repo does not provide
defaults, whereas ours does, and the canton one takes priority on the
class path.
This PR stops daml-script using these defaults. We should align our
scalapb to canton, but this turned out to be more complex than expected.

- https://github.com/digital-asset/daml/pull/22217

The goal is to use the scalapb options from Canton to ensure we generate
the same source files.

---------

Co-authored-by: Samuel Williams <samuel.williams@digitalasset.com>