[Backport #22239] Fix Daml Studio DocumentSymbols error (#22244)
Backport of #22239

Relevant GHCIDE PR here:
https://github.com/digital-asset/daml-ghcide/pull/39

The GHCIDE Pr fixes the language server from giving back bad responses,
however this doesn't fix the IDE for 3.3
This PR also includes a patch mechanism in the multi-ide to fix
responses as they come from the language servers, to hopefully minimise
the times this error is hit.

I tried to do this in the language client VSCode extension, but it
doesn't allow us to manipulate the response before it sends it through
VSCode's validation, which is what's throwing the error. (VSCode
LanguageClient.middleware isn't strong enough)
We could avoid this by writing middleware that replicates the call to
the language server, but thats quite messy.