generate lf hs versions based on bzl versions (#22182)
We generate the lf versions and variables assigned to those versions
(such as DEFAULT_VERSION) based on variables in .bzl files. Because
this, we take a step towards a single source of truth: bazel and haskell
now always agree. Later, we include features, stable packages, and make
scala counterparts. This PR involves the haskell side only