move contract ID version parameter from Engine config to interpretati… (#22033)
…on methods

We are likely not going to perform a big-bang migration from Contract ID
V1 to V2 in Canton. Accordingly, a Canton participant must support to
produce V1 contract IDs for one synchronizer and V2s for others. Yet,
the engine is shared across all synchronizer connections (to save
memory, in particular for the package cache). Accordingly, we cannot
specify the desired contract IDs in the engine config any more, but this
has to be moved to the individual calls made to the engine.