add test cases demonstrating key collisions (#22282)
Keys are currently broken in the presence of force-vetted invalid
upgrades (or missing incompatible creation package).

This PR adds test cases which should result in "key not found" errors
but instead either succeed or fail with an upgrade validation error.
These tests are therefore marked as broken and should "fail to be
broken" once I fix keys.