[backport] Reject tycons in variants and enums in importValue  (#22306)
Backport of https://github.com/digital-asset/daml/pull/22291

Removes tycon IDs in variants an enums in
`Value.nonVerboseWithoutTrailingNones`, which is used by the IDE ledger
to normalize the values (including keys) it gives back to the engine as
a response to `ResultNeedContract`.

That lead me to wonder why this hadn't been caught by any of our IDE
ledger tests since `speedy.ValueTranslator` should reject such values.
Turns out it didn't and is too lenient. This PR fixes this.