generate hs stable packages list (#22252)
This PR adds the functionality of generating a list of stable package
ids based on the haskell-compiled packages. Based on this generated
list, the enoder (which the part that generates the haskell-compiled
packages depends on) can decide if a package is stable based on this
list only, therefore breaking the circular dependency. Updating this
list requires bootstrapping, see
//compiler/damlc/stable-packages/stable-packages.yaml for more info