[3.4] Exclude Scalapb options files from published proto artifact (#22371)
Publishing the `package.proto` file, which contains the Scalapb options,
is inconvenient because:
- It forces our users to add `scalapb-runtime` dep in their classpath,
even if they just want to generate the Java classes
- It forces our users to use the same Scalapb options as us, which is
likely not what they need.

The `package.proto` file was not included in the proto JAR before 3.4.
It has caused issues in transcode, scribe and cn-quickstart since then.