Add logging of canton to debug flaky assistant integration tests (#22430)
This PR enables verbose debug information if it fails whilst running
//daml-assistant/integration-tests:integration-tests. In general we
prefer to not dump so much to stderr but canton exits with error code
117 sometimes causing a flake and it is not reproducible locally.
Therefore we log everything and wait until it flakes such that we get
the information this way.

We log only in case of errors, so for normal runs the log should be
unaffected.