process changes to LATEST from a different branch than main (#22456)
During the repo reshuffle period, we're expecting that `main` may be red
or hard to merge to for a couple of days. We however don't want to block
releases during that period, so we're moving the processing of changes
to the LATEST file to the `release-trigger` branch.

Once the repo reshuffle is complete, we will roll back this change and
copy the LATEST file from the `release-trigger` branch to the `main`
branch.