Add ScriptEngine, working on Values. Update Daml-Script to use this over speedy machine. (#22452)
This PR removes as many references to speedy or SExprs/SValues from
daml-script and adjacent as possible.
Daml script now inspects and constructs Values, rather than SValues.
Values are missing the 3 following constructs that daml-script needs:
- Closures
- Any
- TypeRep
These 3 have been added to Values restricted by a type var, such that
only an `ExtendedValue` can contain them.
The Closure type contains an SPAP from speedy, but is treated as a
blackbox to avoid dependence directly on internals.

The ScriptEngine class has been added as scripts entry point to the
speedy machine.
It takes CompiledPackages and something to run (be that an Identifier,
ExceptionMessage or Interface View, along with relevant ExtendedValues)
and returns an ExtendedValue. It abstracts away the `Result` type to
avoid dealing with Questions.
The ScriptEngine also contains the logic to add an unsafeCoerce to the
engine, which daml-script depends on, by manipulating CompiledPackages.

I was not yet able to remove CompiledPackages manipulation from the
script-service. This may be achievable by extending the
MutableCompiledPackages interface.

Points of note:
- ScriptEngine re-exports some Machine definitions, such as TraceLog and
Warning log, and their builders
- Values are Coerceable to ExtendedValues, but not backwards. GenValue
exposes a `castExtendedValue` function to do this, which will fail on
Closures, Any or TypeRep
- Importing/translation of ExtendedValues into SValues is handled by a
new, simplified ExtendedValueTranslator. This translator does not ensure
type safety as this isn't possible with arbitrary Closures. This
translator expects verbose/enriched values, so we use the enricher in
several places across daml-script to ensure this. The enricher has the
added benefit of applying SCU transformation, which were previously
handled by the Value Translator (that daml-script no longer needs)
- We often use matches like `ValueRecord(_, List((_, val1), (_, val2)))`
to ignore the field names. We may consider a destructor that discards
these for us, to make the logic clearer.
- I've removed the duplicate/unused InterpretationError and Question
classes from daml-script, which are both now provided by Free.scala
- Daml Script previously used the engine to convert legacy exceptions to
failures status, by calling SBThrow. We can no longer do this, so this
conversion logic is replicated in daml-script (still calls the engine
for exception message calculator). It may be possible to refactor this
logic in the engine to make it more directly callable (via
ScriptEngine), by having the engine spinup a new engine, rather than
relying on a special Kontinuation.
- Script still depends on speedy for SError and Pretty (for printing
said SError). This could be re-exported by ScriptEngine or moved
entirely.
- Profile was removed from daml-script, as it was unused.